refactor: finish cart
