+++ utils/cart-tool-kit.js
@@ -5,6 +5,7 @@ const { Cart, CartItem } = require('../db/models')
 const { PREFIX_CART_KEY, PREFIX_CARTITEM_KEY } = require('../config/app').cache.CART
 const { RedisToolKit } = require('../utils/redis-tool-kit')
 const { AuthToolKit } = require('../utils/auth-tool-kit')
+const { ProductToolKit } = require('../utils/product-tool-kit')
 const { ParameterValidationKit } = require('./parameter-validation-kit')
 const { code } = require('../config/result-status-table').errorTable
 
@@ -241,6 +242,106 @@ class CartToolKit {
       }
       return template
     }
+  }
+
+  static async postCartItemsValidate(req) {
+    const { productId } = req.body
+    const redisClient = req.app.locals.redisClient
+    const productKey = `product:${productId}`
+    // check whether parameter syntax is correct
+    let result = {}
+    // const { error, result } = CartToolKit.cartItemSyntaxValidate(req)
+    const syntaxValidation = CartToolKit.cartItemSyntaxValidate(req)
+    if (syntaxValidation.error) {
+      const { result } = syntaxValidation
+      return { error: true, result }
+    }
+    // check whether product exists in product
+    const product = await redisClient.hgetall(productKey)
+    // nothing
+    if (!Object.keys(product).length) {
+      result = { code: code.NOTFOUND, data: null, message: '找不到對應項目' }
+      return { error: true, result }
+    }
+
+    // I've found that
+    // check whether the stock is enough
+    const { cartId } = req.session
+    const cartKey = `${PREFIX_CARTITEM_KEY}:${cartId}:${productId}`
+    const cartItem = await redisClient.hgetall(cartKey)
+    const isExistCart = Boolean(Object.keys(cartItem).length) && Boolean(Number(cartItem.quantity))
+
+    const quantity = isExistCart ? Number(cartItem.quantity) + 1 : 1
+
+    const cartHashMap = {}
+    cartHashMap[productId] = quantity
+
+    const stockHashMap = await ProductToolKit.getStock(productId, redisClient)
+    const { soldOut, notEnough } = await ProductToolKit.checkStockStatus(cartHashMap, stockHashMap)
+    const stockError = Boolean(soldOut.length) || Boolean(notEnough.length)
+
+    // if not enough, just say sorry and return
+    if (stockError) {
+      result = { code: code.BADREQUEST, data: { soldOut, notEnough }, message: '庫存問題' }
+      return { error: true, result }
+    }
+
+    result = { cartHashMap, stockHashMap, cartItem }
+    return { error: false, result }
+  }
+
+  static async putCartItemsValidate(req) {
+    const { isUndefined } = ParameterValidationKit
+    const redisClient = req.app.locals.redisClient
+    const { cartId } = req.session
+    const { items } = req.body
+    const defaultData = isUndefined(items) ? null : JSON.stringify(items)
+    const cart = items
+
+    let result = {}
+    const syntaxValidation = ProductToolKit.quantityHashMapSyntaxValidate(cart)
+
+    if (syntaxValidation.error) {
+      const { result } = syntaxValidation
+      return { error: true, result }
+    }
+
+    const cartHashMap = ProductToolKit.getQuantityHashMap(cart)
+
+    const entries = Object.entries(cartHashMap)
+    const keys = entries.map(([key, _]) => key)
+
+    async function ExistenceTest(keys, cache) {
+      for (const key of keys) {
+        const result = await cache.hgetall(key)
+        if (result.quantity === '0') return false
+        if (!Object.keys(result).length) return false
+      }
+      return true
+    }
+
+    // check whether one of products is not inside the cart
+    const cartKeys = keys.map(item => `${PREFIX_CARTITEM_KEY}:${cartId}:${item}`)
+    const areValidProducts = await ExistenceTest(cartKeys, redisClient)
+
+    if (!areValidProducts) {
+      result = { code: code.NOTFOUND, data: defaultData, message: '購物車內找不到對應項目' }
+      return { error: true, result }
+    }
+
+    //  check whether stock is enough?
+    const { getStock, checkStockStatus } = ProductToolKit
+    const stockHashMap = await getStock(keys, redisClient)
+    const { soldOut, notEnough } = checkStockStatus(cartHashMap, stockHashMap)
+    const stockError = Boolean(soldOut.length) || Boolean(notEnough.length)
+
+    if (stockError) {
+      result = { code: code.BADREQUEST, data: { soldOut, notEnough }, message: '庫存問題' }
+      return { error: true, result }
+    }
+
+    result = { cartHashMap, stockHashMap }
+    return { error: false, result }
   }
 }
